https://github.com/anastyadru/SkyRoads2.git

+++ Assets/Scripts/ScoreManager.cs
@@ -2,13 +2,14 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
+using UnityEngine.SceneManagement;
 
 public class ScoreManager : MonoBehaviour
 {
     [SerializeField] private Text HighScoreText;
     [SerializeField] private Text ScoreText;
 
-    public float score;
+    public int score;
     public int highscore;
 
     void Start()
